move src folder to hakai_qc_app
