+++ Dockerfile
@@ -21,4 +21,4 @@ RUN poetry config virtualenvs.create false \
 
 EXPOSE 8050
 
-CMD ["python", "./src/app.py"]
+CMD ["python", "./hakai_qc_app/app.py"]
