+++ hakai_qc_app/assets/figure_presets.json
