+++ hakai_qc_app/assets/main.css
