+++ hakai_qc_app/selection.py
@@ -20,7 +20,7 @@ from hakai_qc.nutrients import nutrient_variables, run_nutrient_qc
 from hakai_qc.qc import update_dataframe
 
 # from pages.nutrients import get_flag_var
-from utils import load_config
+from hakai_qc_app.utils import load_config
 
 config = load_config()
 variables_flag_mapping = {"no2_no3_um": "no2_no3_flag"}
@@ -436,7 +436,7 @@ def generate_qc_table_style(data):
     logger.debug("Flag columns: {}", flag_columns)
     color_conditional = (
         {
-            "if": {"column_id": col, "filter_query":  "{%s} = '%s'" % (col, flag)},
+            "if": {"column_id": col, "filter_query": "{%s} = '%s'" % (col, flag)},
             "backgroundColor": flag_color,
             "color": "white",
         }
@@ -445,7 +445,7 @@ def generate_qc_table_style(data):
     )
     blank_conditional = (
         {
-            "if": {"column_id": col, "filter_query":  "{%s} = ''"  % col},
+            "if": {"column_id": col, "filter_query": "{%s} = ''" % col},
             "backgroundColor": "light_grey",
             "color": "black",
         }
