+++ hakai_qc_app/tooltips.py
