+++ makefile
@@ -1,5 +1,5 @@
 run_dash:
-	poetry run python src/app.py
+	poetry run python hakai_qc_app/app.py
 
 clear_log:
 	rm logs/dashboard.log
