+++ tests/test_hakai_auth.py
@@ -1,0 +1,12 @@
+from hakai_qc_app.download_hakai import parse_hakai_token
+import pytest
+
+
+@pytest.mark.parametrize(
+    "token",
+    [
+        "token_type=Bearer&access_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjExNjMxNzQzNDYzMTA5NzUwOTczOSIsImVtYWlsIjoiamVzc3kuYmFycmV0dGVAaGFrYWkub3JnIiwidmVyaWZpZWRfZW1haWwiOnRydWUsIm5hbWUiOiJKZXNzeSBCYXJyZXR0ZSIsImdpdmVuX25hbWUiOiJKZXNzeSIsImZhbWlseV9uYW1lIjoiQmFycmV0dGUiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDMuZ29vZ2xldXNlcmNvbnRlbnQuY29tL2EvQUNnOG9jSUtHQjVZdFkzODYxSFZGczluSHVzbnpydXJLVUFibTdCaEJYSUhJdnNQZFE9czk2LWMiLCJsb2NhbGUiOiJlbiIsImhkIjoiaGFrYWkub3JnIiwiZ3JvdXBzIjpbImplc3N5LmJhcnJldHRlQGhha2FpLm9yZyIsImFsZ2FlZXhwbG9yZXJAaGFrYWkub3JnIiwiY2lvb3MucGFjaWZpY0BoYWthaS5vcmciLCJkYXRhLm1vYmlsaXphdGlvbkBoYWthaS5vcmciLCJlaW1zLmF1dGguaGFrYWlAaGFrYWkub3JnIiwiZWltcy5hdXRoLm5hdHVyZS50cnVzdEBoYWthaS5vcmciLCJlaW1zLmF1dGgucGFya3MuY2FuYWRhQGhha2FpLm9yZyIsImVpbXMuYXV0aC5zZmNAaGFrYWkub3JnIiwiZWltcy5hdXRoLnV3dEBoYWthaS5vcmciLCJlaW1zLmluc3RydW1lbnRhdGlvbi5hZG1pbkBoYWthaS5vcmciLCJlaW1zLmluc3RydW1lbnRhdGlvbkBoYWthaS5vcmciLCJlaW1zLm9jZWFub2dyYXBoeS5lZGl0b3JzQGhha2FpLm9yZyIsImVpbXMub2NlYW5vZ3JhcGh5LnFjQGhha2FpLm9yZyIsImVpbXMub2NlYW5vZ3JhcGh5LnVzZXJzQGhha2FpLm9yZyIsImVyZGRhcC1ub3RpZmljYXRpb25zQGhha2FpLm9yZyIsIm1hcmluZS13b3JraW5nLWdyb3VwQGhha2FpLm9yZyIsIm9jZWFuLW9ic2VydmluZy1wbGFubmluZy1ncm91cEBoYWthaS5vcmciLCJvY2Vhbm9ncmFwaHlAaGFrYWkub3JnIiwicXUzOV9vYV9wYXBlckBoYWthaS5vcmciLCJzbi5hbGVydHNAaGFrYWkub3JnIiwic3RhZmZAaGFrYWkub3JnIiwidGVjaC50ZWFtQGhha2FpLm9yZyIsImVpbXMuZWRpdG9yc0BoYWthaS5vcmciLCJlaW1zLmluc3RydW1lbnRhdGlvbi51c2Vyc0BoYWthaS5vcmciLCJlaW1zLnVzZXJzQGhha2FpLm9yZyJdLCJ3b3JrYXJlYXMiOlsiR1dBSUkgSEFBTkFTIiwiRVNUVUFSWSIsIkNMQVlPUVVPVCIsIlNLRUVOQSIsIkNBTFZFUlQiLCJKT0hOU1RPTkUgU1RSQUlUIiwiUVVBRFJBIiwiTE9XRVIgTUFJTkxBTkQiXSwib3JnYW5pemF0aW9ucyI6WyJQQVJLUyBDQU5BREEiLCJOQVRVUkUgVFJVU1QiLCJVV1QiLCJTRkMiLCJIQUtBSSJdLCJpYXQiOjE3MDA2NjY0MTMsImV4cCI6MTcwMDcwOTYxMywiaXNzIjoiaGFrYWktYXBpIn0.4PKa-mzNztFVa74szepjw83o889eK_f9WhZQPt70a2c&expires_at=1700709613"
+    ],
+)
+def test_token_parser(token):
+    parse_hakai_token(token)
